Add additional state for inactive customers
We need a way for the game to know when a customer is no longer in the game. The easiest way would be to delete the customer from the 'customers' map and from scene.drawables. However, as an easy workaround right now, I am adding an extra state called "Inactive", which is for inactive customers. After a customer has entered the inactive state, it will do nothing